# internal/adapter/in/http/handler_tokens.go
package http

import (
	"encoding/json"
	"errors"
	"net/http"
	"time"

	"phone-tokens/internal/model"
	"phone-tokens/internal/service/tokens"
)

type TokenHandler struct {
	service tokens.Service
}

func NewTokenHandler(service tokens.Service) *TokenHandler {
	return &TokenHandler{service: service}
}

type createTokenRequest struct {
	Name        string                  `json:"name"`
	Permissions []model.TokenPermission `json:"permissions"`
	TTLSeconds  int64                   `json:"ttl_seconds"`
}

type tokenResponse struct {
	ID          string                  `json:"id"`
	Token       string                  `json:"token"`
	Name        string                  `json:"name"`
	Permissions []model.TokenPermission `json:"permissions"`
	Status      model.TokenStatus       `json:"status"`
	ExpiresAt   string                  `json:"expires_at"`
}

// CreateToken godoc
// @Summary Issue a persistent token
// @Security BearerAuth
// @Description Issues a persistent token for the authenticated user
// @Tags Token
// @Accept json
// @Produce json
// @Param request body createTokenRequest true "Token request payload"
// @Success 201 {object} tokenResponse "Token successfully issued"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /api/v1/tokens [post]
func (h *TokenHandler) CreateToken(w http.ResponseWriter, r *http.Request) {
	var req createTokenRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}
	if req.TTLSeconds <= 0 {
		http.Error(w, "ttl_seconds must be greater than zero", http.StatusBadRequest)
		return
	}

	claims, ok := r.Context().Value(userContextKey).(*UserClaims)
	if !ok || claims.UserID == "" {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}

	token, err := h.service.Issue(r.Context(), tokens.IssueInput{
		UserID:      claims.UserID,
		Name:        req.Name,
		Permissions: req.Permissions,
		TTLSeconds:  req.TTLSeconds,
	})
	if err != nil {
		if isValidationError(err) {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		http.Error(w, "internal error", http.StatusInternalServerError)
		return
	}

	writeJSON(w, http.StatusCreated, toTokenResponse(token))
}

type updateTokenRequest struct {
	TTLSeconds int64 `json:"ttl_seconds"`
}

// UpdateTokenTTL godoc
// @Summary Update token expiration
// @Security BearerAuth
// @Description Updates token expiration time for the authenticated user
// @Tags Token
// @Accept json
// @Produce json
// @Param tokenID path string true "Token ID"
// @Param request body updateTokenRequest true "Update TTL request payload"
// @Success 200 {object} tokenResponse "Token successfully updated"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 403 {object} map[string]string "Forbidden"
// @Failure 404 {object} map[string]string "Token not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /api/v1/tokens/{tokenID} [patch]
func (h *TokenHandler) UpdateTokenTTL(w http.ResponseWriter, r *http.Request) {
	var req updateTokenRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}
	if req.TTLSeconds <= 0 {
		http.Error(w, "ttl_seconds must be greater than zero", http.StatusBadRequest)
		return
	}

	tokenID := r.PathValue("tokenID")
	if tokenID == "" {
		http.Error(w, "token id is required", http.StatusBadRequest)
		return
	}

	claims, ok := r.Context().Value(userContextKey).(*UserClaims)
	if !ok || claims.UserID == "" {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}

	token, err := h.service.UpdateTTL(r.Context(), claims.UserID, tokenID, req.TTLSeconds)
	if err != nil {
		switch err {
		case tokens.ErrForbidden:
			http.Error(w, err.Error(), http.StatusForbidden)
		case tokens.ErrNotFound:
			http.Error(w, err.Error(), http.StatusNotFound)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}

	writeJSON(w, http.StatusOK, toTokenResponse(token))
}

// DeleteToken godoc
// @Summary Delete a token
// @Security BearerAuth
// @Description Removes a token for the authenticated user
// @Tags Token
// @Param tokenID path string true "Token ID"
// @Success 204 "No Content"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 403 {object} map[string]string "Forbidden"
// @Failure 404 {object} map[string]string "Token not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /api/v1/tokens/{tokenID} [delete]
func (h *TokenHandler) DeleteToken(w http.ResponseWriter, r *http.Request) {
	tokenID := r.PathValue("tokenID")
	if tokenID == "" {
		http.Error(w, "token id is required", http.StatusBadRequest)
		return
	}

	claims, ok := r.Context().Value(userContextKey).(*UserClaims)
	if !ok || claims.UserID == "" {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}

	err := h.service.Delete(r.Context(), claims.UserID, tokenID)
	if err != nil {
		switch err {
		case tokens.ErrForbidden:
			http.Error(w, err.Error(), http.StatusForbidden)
		case tokens.ErrNotFound:
			http.Error(w, err.Error(), http.StatusNotFound)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// FreezeToken godoc
// @Summary Freeze a token
// @Security BearerAuth
// @Description Sets token status to frozen
// @Tags Token
// @Param tokenID path string true "Token ID"
// @Success 200 {object} tokenResponse "Token successfully frozen"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 403 {object} map[string]string "Forbidden"
// @Failure 404 {object} map[string]string "Token not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /api/v1/tokens/{tokenID}/freeze [patch]
func (h *TokenHandler) FreezeToken(w http.ResponseWriter, r *http.Request) {
	h.changeStatus(w, r, model.TokenStatusFrozen)
}

// UnfreezeToken godoc
// @Summary Unfreeze a token
// @Security BearerAuth
// @Description Sets token status to active
// @Tags Token
// @Param tokenID path string true "Token ID"
// @Success 200 {object} tokenResponse "Token successfully unfrozen"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 403 {object} map[string]string "Forbidden"
// @Failure 404 {object} map[string]string "Token not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /api/v1/tokens/{tokenID}/unfreeze [patch]
func (h *TokenHandler) UnfreezeToken(w http.ResponseWriter, r *http.Request) {
	h.changeStatus(w, r, model.TokenStatusActive)
}

func (h *TokenHandler) changeStatus(w http.ResponseWriter, r *http.Request, status model.TokenStatus) {
	tokenID := r.PathValue("tokenID")
	if tokenID == "" {
		http.Error(w, "token id is required", http.StatusBadRequest)
		return
	}

	claims, ok := r.Context().Value(userContextKey).(*UserClaims)
	if !ok || claims.UserID == "" {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}

	token, err := h.service.SetStatus(r.Context(), claims.UserID, tokenID, status)
	if err != nil {
		switch err {
		case tokens.ErrForbidden:
			http.Error(w, err.Error(), http.StatusForbidden)
		case tokens.ErrNotFound:
			http.Error(w, err.Error(), http.StatusNotFound)
		default:
			if isValidationError(err) {
				http.Error(w, err.Error(), http.StatusBadRequest)
				return
			}
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}

	writeJSON(w, http.StatusOK, toTokenResponse(token))
}

func toTokenResponse(token *model.UserToken) tokenResponse {
	return tokenResponse{
		ID:          token.ID,
		Token:       token.Token,
		Name:        token.Name,
		Permissions: token.Permissions,
		Status:      token.Status,
		ExpiresAt:   token.ExpiresAt.Format(time.RFC3339),
	}
}

func isValidationError(err error) bool {
	return errors.Is(err, tokens.ErrInvalidPermission) || errors.Is(err, tokens.ErrInvalidStatus)
}